refactor - 네이밍 수정.
