+++ app/src/androidTest/java/nextstep/payments/component/CardNumberTextFieldTest.kt
@@ -7,25 +7,25 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.requestFocus
-import nextstep.payments.ui.newcard.component.CardExpiredDateTextFiled
+import nextstep.payments.ui.newcard.component.CardNumberTextField
 import org.junit.Rule
 import org.junit.Test
 
-class CartExpiredDateTextFiledTest {
+class CardNumberTextFieldTest {
+
     @get:Rule
     val composeTestRule = createComposeRule()
-
 
     @Test
     fun `초기화면에_라벨이_보여야한다`() {
         val input = mutableStateOf("")
 
         composeTestRule.setContent {
-            CardExpiredDateTextFiled(input.value, onValueChange = { input.value = it })
+            CardNumberTextField(input.value, onValueChange = { input.value = it })
         }
 
         composeTestRule
-            .onNodeWithText("만료일")
+            .onNodeWithText("카드 번호")
             .assertIsDisplayed()
     }
 
@@ -34,15 +34,15 @@ class CartExpiredDateTextFiledTest {
         val input = mutableStateOf("")
 
         composeTestRule.setContent {
-            CardExpiredDateTextFiled(input.value, onValueChange = { input.value = it })
+            CardNumberTextField(input.value, onValueChange = { input.value = it })
         }
 
         composeTestRule
-            .onNodeWithContentDescription("CardExpiredDateTextFiled")
+            .onNodeWithContentDescription("CardNumberTextField")
             .requestFocus()
 
         composeTestRule
-            .onNodeWithText("MM / YY")
+            .onNodeWithText("0000 – 0000 – 0000 – 0000")
             .assertIsDisplayed()
     }
 
@@ -51,22 +51,23 @@ class CartExpiredDateTextFiledTest {
         val input = mutableStateOf("12")
 
         composeTestRule.setContent {
-            CardExpiredDateTextFiled(input.value, onValueChange = { input.value = it })
+            CardNumberTextField(input.value, onValueChange = { input.value = it })
         }
         composeTestRule
-            .onNodeWithText("MM / YY")
+            .onNodeWithText("0000 – 0000 – 0000 – 0000")
             .assertIsNotDisplayed()
     }
 
     @Test
     fun `입력이_포멧에_맞게_올바르게_나와야_한다`() {
-        val input = mutableStateOf("0225")
+        val input = mutableStateOf("1234123412341234")
 
         composeTestRule.setContent {
-            CardExpiredDateTextFiled(input.value, onValueChange = { input.value = it })
+            CardNumberTextField(input.value, onValueChange = { input.value = it })
         }
+
         composeTestRule
-            .onNodeWithText("02 / 25")
+            .onNodeWithText("1234 – 1234 – 1234 – 1234")
             .assertIsDisplayed()
     }
 }
